updated configs
