+++ .isort.cfg
@@ -1,4 +1,5 @@
 [settings]
+profile = black
 known_third_party =
     invoke
     pylint
