+++ src/cdkutils/persistent_config.py
@@ -9,7 +9,6 @@ import boto3
 from aws_cdk import aws_ssm as ssm
 from aws_cdk import core as cdk
 
-from cdkutils.config import SsmConfig
 from cdkutils.errors import AttributeNotFoundException, SecretCreationException
 
 if TYPE_CHECKING:
@@ -20,6 +19,59 @@ else:
     SecretsManagerClient = object
 
 _LOGGER = logging.getLogger(__name__)
+
+
+class SsmConfig:
+    """Represents SSM config details"""
+
+    def __init__(
+        self,
+        namespace: Optional[str] = None,
+        config_id: Optional[str] = None,
+        org_name: Optional[str] = None,
+        cdk_scope: Optional[cdk.Construct] = None,
+    ) -> None:
+
+        if namespace is None and cdk_scope:
+            namespace = cdk_scope.node.try_get_context("SsmNamespace")
+
+        if namespace is None:  # i.e. if namespace is *still* None
+            raise ValueError("a namespace must be provided, either via a parameter or CDK context variable")
+        self.namespace = namespace
+
+        if cdk_scope is not None and config_id is None:
+            config_id = cdk_scope.node.try_get_context("SsmConfigId")
+        self.config_id = "default" if config_id is None else config_id
+
+        if cdk_scope is not None and org_name is None:
+            org_name = cdk_scope.node.try_get_context("SsmConfigOrg")
+        self.org_name = "metoffice" if org_name is None else org_name
+
+    def __eq__(self, other: Any) -> bool:
+        return (
+            isinstance(other, SsmConfig)
+            and self.config_id == other.config_id
+            and self.org_name == other.org_name
+            and self.namespace == other.namespace
+        )
+
+    def __repr__(self) -> str:
+        return f"SsmConfig({self.namespace!r}, {self.config_id!r}, {self.org_name!r})"
+
+    def __str__(self) -> str:
+        return self.path_prefix
+
+    @property
+    def path_prefix(self) -> str:
+        """Return the path prefix common to all config entries relating to this config"""
+        return f"/{self.org_name}/{self.namespace}/{self.config_id}"
+
+    def get_full_path(self, partial_path: str) -> str:
+        """
+        Return full SSM path for the given partial_path, which is the same as prepending the path_prefix to partial_path
+        """
+        return f"{self.path_prefix}/{partial_path}"
+
 
 @dataclass
 class PersistedAttribute:
@@ -130,8 +182,8 @@ class PersistedConfig(ABC):
                     else:
                         val = ssm_client.get_parameter(Name=full_path)["Parameter"]["Value"]
                 except (
-                        ssm_client.exceptions.ParameterNotFound,
-                        secret_mgr_client.exceptions.ResourceNotFoundException,
+                    ssm_client.exceptions.ParameterNotFound,
+                    secret_mgr_client.exceptions.ResourceNotFoundException,
                 ) as exc:
                     default_value = inspect.signature(cls.__init__).parameters[persisted_attr.attribute].default
                     if default_value and default_value != inspect.Parameter.empty:
@@ -271,18 +323,17 @@ class PersistedConfig(ABC):
             getattr(self, subconfig).create_secrets(boto_session)
 
     @final
-    def update_existing_secret(
-        self, name: str, secret_string: str, sm_client: SecretsManagerClient
-    ) -> None:
+    @staticmethod
+    def update_existing_secret(name: str, secret_string: str, sm_client: SecretsManagerClient) -> None:
 
         try:
             sm_client.update_secret(SecretId=name, SecretString=secret_string)
             _LOGGER.info(f"Updated {name} in Secret Manager")
 
         except (
-                sm_client.exceptions.InvalidRequestException,
-                sm_client.exceptions.InvalidParameterException,
-                sm_client.exceptions.LimitExceededException,
+            sm_client.exceptions.InvalidRequestException,
+            sm_client.exceptions.InvalidParameterException,
+            sm_client.exceptions.LimitExceededException,
         ) as exc:
             msg = f"Error creating {name}: {exc}"
             _LOGGER.error(msg)
@@ -299,7 +350,7 @@ class PersistedConfig(ABC):
         for mapping in self._get_persisted_attributes():
             if mapping.use_secrets_manager:
                 with contextlib.suppress(
-                        sm_client.exceptions.ResourceNotFoundException, sm_client.exceptions.InvalidRequestException
+                    sm_client.exceptions.ResourceNotFoundException, sm_client.exceptions.InvalidRequestException
                 ):
                     name = ssm_config.get_full_path(mapping.aws_partial_path)
                     sm_client.delete_secret(SecretId=name, ForceDeleteWithoutRecovery=no_recovery)
