+++ tox.ini
@@ -31,7 +31,6 @@ commands =
     coverage report
     coverage xml
 
-
 [testenv:pylint]
 deps =
     -r requirements/requirements-linting.txt
